+++ internal/modules/checksum.go
@@ -32,7 +32,7 @@ type Checksum struct {
 // nolint: gochecknoinits
 func init() {
 	modules.RegisterModule(&modules.ModuleRegistration{
-		Stage:   "archive",
+		Stage:   "build",
 		Type:    "checksum",
 		Factory: NewChecksum,
 	})