+++ internal/modules/tar.go
@@ -48,7 +48,7 @@ type (
 // nolint: gochecknoinits
 func init() {
 	modules.RegisterModule(&modules.ModuleRegistration{
-		Stage:   "archive",
+		Stage:   "build",
 		Type:    "tar",
 		Factory: NewTar,
 	})